libcutils: test: increase use of ashmem_valid(fd)
There may be evidence of ashmem_valid(fd) reporting that the file
descriptor is an ashmem node.  Increase testing of ashmem_valid(fd),
reporting that the node _is_ ashmem, to inspire confidence in the
positive result.  Scan all file descriptors in the system, and for
those that pass ashmem_valid, get a non-zero size reference back.

Some clang-format-isms applied.

Test: libcutils-test --gtest_filter=AshmemTest.*
Bug: 72021458
Change-Id: I77d746b57a89a6afa1b829dddfdc4dd319f6b684

Former-commit-id: 4eae47043869176dddda14d7d27e73d9c2bbb876 [formerly cba6b78c331cedbaf47e76e34828110b554bfcb5] [formerly cba6b78c331cedbaf47e76e34828110b554bfcb5 [formerly 276758dab8e24b29be0abb1b6bdadd7c897e8701]]
Former-commit-id: 2f85df9209c644c346762d00c8ee7823bec116b6
Former-commit-id: 21fa7e882285c82f1f8f035c45deab958639810f