Update stable_properties.h
ro.boot.* are from kernel cmdline, and kernel is usually owned by SoC
vendor or ODM.
So those properties should be allowed as action triggers of
vendor/odm init scripts.

Additionally the state of mediadrm (/system/bin/mediadrmserver) should
be used to operate the state of drm HAL.
So init.svc.mediadrm should be whitelisted as an action trigger as well.

Bug: 36796459
Test: tested with walleye
Change-Id: Ic9f68162c577cc190f193063988ad04e42478e6e

Former-commit-id: 88444aca8c97dd91f7406bdc3b8e2e3b009c63ed [formerly 015f1a90c5bfbcf07cb4ccb2611ab241ccb1637c] [formerly 015f1a90c5bfbcf07cb4ccb2611ab241ccb1637c [formerly cf2da29d66d641d3b2f27de886403b28804cbe7b]]
Former-commit-id: b1d99138577a75dd770428d8ff32b54aeb190ee8
Former-commit-id: 6c4afec2acb08111a14cff041227ab2be1592aa1